+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.5] - 2026-06-03
+
+### Docs
+- Update README.md
+- Update docs/clients/aider-docker-autoloop.md
+
+### Test
+- Update tests/test_sllm.py
+
+### Other
+- Update uv.lock
+
 ## [0.1.4] - 2026-06-03
 
 ### Docs